quantum_maze/requirements.txt
pygame-ce>=2.3.0
numpy>=1.24.0
pytweening>=1.0.4
